Serde-encode info data instead of generating code.
This encodes the register tables parsed from the Mesa sources into a
serialized data structure, which is then embedded in the rodata of the
vsc7448-info crate and deserialized on-demand.

Currently I'm using postcard, the same codec Bryan uses in portions of
Humility, because it's faster and smaller than bincode. I experimented
with using human-readable formats like RON, but, they slow.

This reduces release compile time for the vsc7448-info crate on my Ryzen
7 laptop from 1m42s to ~5s. Runtime parsing of the serialized tables
takes about 5ms (in a release build, 46ms in a debug build).